// src/types/user.ts
export interface UserInfo {
	userSerialId?: number;
	firstName?: string;
	lastName?: string;
	avatar?: string;
	usedDisk?: number;
	availableDisk?: number;
	joinAt?: string;
	isActive?: boolean;
	accountSerialId?: number;
	phoneNumber?: number;
	address?: string;
	isAutoBuildGnn?: boolean;
	isAutoDetectMaterial?: boolean;
	isAutoBuildAnn?: boolean;
	isUseChatBotAi?: boolean;
}